+++ app/src/main/java/com/example/android_accountbook_13/presenter/AccountBookApp.kt
@@ -4,9 +4,6 @@ import androidx.compose.material.Scaffold
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.tooling.preview.Preview
 import com.example.android_accountbook_13.presenter.component.BottomAppBar
-import com.example.android_accountbook_13.presenter.navigation.AccountBookNavHost
-import com.example.android_accountbook_13.presenter.navigation.History
-import com.example.android_accountbook_13.presenter.navigation.navigateSingleTopTo
 import com.example.android_accountbook_13.ui.theme.AccountBookTheme
 
 @Composable
